refactor: events.js 속 객체를 도메인으로 분리
