--- src/constants/constants.js
@@ -1,1 +1,0 @@
-export const ORDER_REGEX = /^([ㄱ-ㅎㅏ-ㅣ가-힣]+-([1-9]|1\d|20),?)+$/;
