+++ src/domain/Events.js
@@ -1,25 +1,5 @@
-import { DESSERTS, MAIN_COURSES } from './menus.js';
-
-export const EVENT_YEAR = 2023;
-export const EVENT_MONTH = 12;
-export const EVENT_PERIOD = Object.freeze({
-  startDate: 1,
-  endDate: 31,
-});
-
-export const BENEFIT_TYPE = Object.freeze({
-  totalDiscount: 'totalDiscount',
-  menuDiscount: 'menuDiscount',
-  gift: 'gift',
-});
-
-export const EVENT_NAMES = Object.freeze({
-  CHRISTMAS_D_DAY: '크리스마스 디데이 할인',
-  WEEKDAY: '평일 할인',
-  WEEKEND: '주말 할인',
-  SPECIAL: '특별 할인',
-  GIFT: '증정 이벤트',
-});
+import { BENEFIT_TYPE, EVENT_NAMES } from '../constants/event.js';
+import { DESSERTS, MAIN_COURSES } from '../constants/menus.js';
 
 export const CHRISTMAS_D_DAY = {
   type: BENEFIT_TYPE.totalDiscount,
