+++ src/services/ReservationController.js
@@ -3,7 +3,6 @@ import OutputView from '../views/OutputView.js';
 import EventPlanner from './EventPlanner.js';
 
 import MESSAGES from '../constants/messages.js';
-import { ORDER_REGEX } from '../constants/constants.js';
 
 import throwError from '../utils/throwError.js';
 import handleException from '../utils/handleException.js';
@@ -82,6 +81,8 @@ class ReservationController {
   }
 
   #validateOrdersFormat(orders) {
+    const ORDER_REGEX = /^([ㄱ-ㅎㅏ-ㅣ가-힣]+-([1-9]|1\d|20),?)+$/;
+
     if (!isValidFormat(ORDER_REGEX, orders))
       throwError(MESSAGES.errors.invalidOrders);
   }
